Polish the NAC module and drop the redundant bundle threading
Generate the 27-image BZ search space instead of tabulating it (origin kept
first so folding ties still fall back to the unshifted point) and remove the
section banners.

Delete the precomputed-bundle threading through Phonons and the
nac_precomputed parameter of HarmonicWithQ: SecondOrder.get_nac_precomputed
already caches per BvK matrix on the force-constant object, so every
HarmonicWithQ sharing those force constants reuses one bundle without any
plumbing. The threading predated that cache and had no remaining callers.

Fold the QE dipole-subtracted convention contract tests into
test_nac_api.py and drop test_nac_qe.py.
